+++ pom.xml
@@ -92,6 +92,12 @@
 			<version>${liquibase.version}</version>
 		</dependency>
 
+		<dependency>
+			<groupId>org.hibernate.validator</groupId>
+			<artifactId>hibernate-validator</artifactId>
+			<version>${hibernate-validator.version}</version>
+		</dependency>
+
 	</dependencies>
 
 	<build>
