+++ src/main/java/com/book_app/dto/CreateBookRequestDto.java
@@ -1,0 +1,21 @@
+package com.book_app.dto;
+
+import java.math.BigDecimal;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+
+@Data
+public class CreateBookRequestDto {
+    @NotNull
+    private String title;
+    @NotNull
+    private String author;
+    @NotNull
+    private String isbn;
+    @NotNull
+    @Min(0)
+    private BigDecimal price;
+    private String description;
+    private String coverImage;
+}
