+++ src/main/java/com/book_app/exception/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package com.book_app.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
