+++ src/main/java/com/book_app/exception/SpecificationProviderException.java
@@ -1,0 +1,7 @@
+package com.book_app.exception;
+
+public class SpecificationProviderException extends RuntimeException {
+    public SpecificationProviderException(String message) {
+        super(message);
+    }
+}
