+++ src/main/java/com/book_app/repository/SpecificationProviderManager.java
@@ -1,0 +1,5 @@
+package com.book_app.repository;
+
+public interface SpecificationProviderManager<T> {
+    SpecificationProvider<T> getSpecificationProvider(String key);
+}
