+++ src/main/java/com/book_app/repository/book/BookRepository.java
@@ -1,0 +1,8 @@
+package com.book_app.repository.book;
+
+import com.book_app.model.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+public interface BookRepository extends JpaRepository<Book, Long>, JpaSpecificationExecutor<Book> {
+}
