--- src/main/java/mate/config/MapperConfig.java
@@ -1,14 +1,0 @@
-package mate.config;
-
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.NullValueCheckStrategy;
-
-@org.mapstruct.MapperConfig(
-        componentModel = "spring",
-        injectionStrategy = InjectionStrategy.CONSTRUCTOR,
-        nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS,
-        implementationPackage = "<PACKAGE_NAME>.impl"
-)
-public class MapperConfig {
-
-}
