--- src/main/java/mate/controller/BookController.java
@@ -1,50 +1,0 @@
-package mate.controller;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import mate.dto.BookDto;
-import mate.dto.BookSearchParametersDto;
-import mate.dto.CreateBookRequestDto;
-import mate.service.BookService;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping(value = "/api/books")
-public class BookController {
-    private final BookService bookService;
-
-    @GetMapping
-    public List<BookDto> findAll() {
-        return bookService.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public BookDto findById(@PathVariable Long id) {
-        return bookService.findById(id);
-    }
-
-    @GetMapping("/search")
-    public List<BookDto> search(BookSearchParametersDto bookSearchParametersDto) {
-        return bookService.search(bookSearchParametersDto);
-    }
-
-    @PostMapping
-    public BookDto save(@RequestBody CreateBookRequestDto createBookRequestDto) {
-        return bookService.save(createBookRequestDto);
-    }
-
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        bookService.deleteById(id);
-    }
-}
