--- src/main/java/mate/dto/BookDto.java
@@ -1,15 +1,0 @@
-package mate.dto;
-
-import java.math.BigDecimal;
-import lombok.Data;
-
-@Data
-public class BookDto {
-    private Long id;
-    private String title;
-    private String author;
-    private String isbn;
-    private BigDecimal price;
-    private String description;
-    private String coverImage;
-}
