--- src/main/java/mate/dto/BookSearchParametersDto.java
@@ -1,8 +1,0 @@
-package mate.dto;
-
-public record BookSearchParametersDto(String[] titles,
-                                      String[] authors,
-                                      String[] isbn,
-                                      String[] price,
-                                      String[] description) {
-}
