--- src/main/java/mate/dto/CreateBookRequestDto.java
@@ -1,14 +1,0 @@
-package mate.dto;
-
-import java.math.BigDecimal;
-import lombok.Data;
-
-@Data
-public class CreateBookRequestDto {
-    private String title;
-    private String author;
-    private String isbn;
-    private BigDecimal price;
-    private String description;
-    private String coverImage;
-}
