--- src/main/java/mate/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package mate.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
