--- src/main/java/mate/exception/SpecificationProviderException.java
@@ -1,7 +1,0 @@
-package mate.exception;
-
-public class SpecificationProviderException extends RuntimeException {
-    public SpecificationProviderException(String message) {
-        super(message);
-    }
-}
