--- src/main/java/mate/repository/SpecificationBuilder.java
@@ -1,8 +1,0 @@
-package mate.repository;
-
-import mate.dto.BookSearchParametersDto;
-import org.springframework.data.jpa.domain.Specification;
-
-public interface SpecificationBuilder<T> {
-    Specification<T> build(BookSearchParametersDto bookSearchParametersDto);
-}
