--- src/main/java/mate/repository/SpecificationProvider.java
@@ -1,9 +1,0 @@
-package mate.repository;
-
-import org.springframework.data.jpa.domain.Specification;
-
-public interface SpecificationProvider<T> {
-    String getKey();
-
-    Specification<T> getSpecification(String[] params);
-}
