--- src/main/java/mate/repository/SpecificationProviderManager.java
@@ -1,5 +1,0 @@
-package mate.repository;
-
-public interface SpecificationProviderManager<T> {
-    SpecificationProvider<T> getSpecificationProvider(String key);
-}
