--- src/main/java/mate/repository/book/BookRepository.java
@@ -1,8 +1,0 @@
-package mate.repository.book;
-
-import mate.model.Book;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
-public interface BookRepository extends JpaRepository<Book, Long>, JpaSpecificationExecutor<Book> {
-}
