+++ src/test/java/com/book_app/DemoApplicationTests.java
@@ -1,0 +1,14 @@
+package com.book_app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
+
