--- src/test/java/mate/DemoApplicationTests.java
@@ -1,14 +1,0 @@
-package mate;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
-
